Sistema para loja fase final, algumas alterações na fase final

+++ SistemaLoja/models/Produto.js
@@ -5,7 +5,7 @@ import connection from "../config/sequelize-config.js"
 const ProdutoMDC = connection.define("produtos", {
     img: Sequelize.STRING,
     nomep: Sequelize.STRING,
-    pre: Sequelize.INTEGER,
+    pre: Sequelize.DOUBLE,
     cat: Sequelize.STRING
 });
 ProdutoMDC.sync({force: false});
